# swagger.yml
openapi: 3.0.0
info:
  title: The Best API Ever
  description: API terbaik bangsa [CommonMark](http://commonmark.org/help/) or HTML.
  version: 0.1.69
servers:
  - url: https://localhost:7195
    description: Optional server description, e.g. Main (production) server
  - url: http://localhost:5262
    description: Optional server description, e.g. Internal staging server for testing
paths:
...
  /auth/login:
    post:
      summary: "Login"
      tags:
        - "Auth"
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                Data: 
                  type: object
                  properties:     
                    Username: 
                      type: string
                    Password:
                      type: string
                    version:
                      type: string
                    DeviceId:
                      type: string
      responses:
        "200": 
          description: "success"
          content:
            application/json:
              schema: 
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      id:
                        type: integer
                      is_registered:
                        type: boolean
                      change:
                        type: string
                      supervisor_code:
                        type: string
                      profile:
                        type: object
                        properties:
                          trainee_id:
                            type: integer
                          trainee_phone:
                            type: string
                          trainee_email:
                            type: string
                          change:
                            type: string
                          name:
                            type: string
                          email:
                            type: string
                          phone:
                            type: string
                          position:
                            type: string
                          level:
                            type: string
                          photo:
                            type: string
                      dealer:
                        type: object
                        properties:
                          id:
                            type: integer
                          name:
                            type: string
                          code:
                            type: string
                          city:
                            type: string
                          dealer:
                            type: string
                          group:
                            type: string
                      analytic_user:
                        type: object
                        properties:
                          dealer_id:
                            type: integer
                          dealer_branch_code:
                            type: string
                          job_position:
                            type: string
                          city:
                            type: string
                      access_token:
                        type: object
                        properties:
                          auth_token:
                            type: string
                          type:
                            type: string
                          expires_at:
                            type: integer
                      force_update:
                        type: boolean
                      is_data_match:
                        type: boolean
                      go_live:
                        type: string
                  success:
                    type: boolean
                  message:
                    type: string
                  origin:
                    type: string
        '400':
          $ref: '#/components/responses/Error400'
        '401':
          $ref: '#/components/responses/Error401'
        '404':
          $ref: '#/components/responses/Error404'
        '500':
          $ref: '#/components/responses/Error500'
  /auth/forgot-password:
    post:
      summary: 'Forgot Password'
      tags:
        - 'Auth'
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                Data: 
                  type: object
                  properties:     
                    Username: 
                      type: string
                    Type:
                      type: string
                    Identifier:
                      type: string
      responses:
        "200": 
          description: "success"
          content:
            application/json:
              schema: 
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  origin:
                    type: string
        '400':
          $ref: '#/components/responses/Error400'
        '401':
          $ref: '#/components/responses/Error401'
        '404':
          $ref: '#/components/responses/Error404'
        '500':
          $ref: '#/components/responses/Error500'
  /auth/register:
    post:
      summary: 'Register'
      tags:
        - Auth
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                Data:
                  type: object
                  properties:
                    Email:
                      type: string
                    Phone:
                      type: string
                    Password:
                      type: string
                    UserName:
                      type: string
                    DeviceId:
                      type: string
                    Version:
                      type: string
      responses:
        '200':
          description: 'success'
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: string
                  success:
                    type: boolean
                  message:
                    type: string
                  origin:
                    type: string
        '400':
          $ref: '#/components/responses/Error400'
        '401':
          $ref: '#/components/responses/Error401'
        '404':
          $ref: '#/components/responses/Error404'
        '500':
          $ref: '#/components/responses/Error500'
  /auth/otp: 
    post: 
      summary: 'OTP'
      tags:
        - Auth
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                Data:
                  type: object
                  properties:
                    UserId:
                      type: integer
                    Pin: 
                      type: string
      responses:
        '200':
          description: success
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: string
                  success:
                    type: boolean
                  message:
                    type: string
                  origin:
                    type: string
        '400':
          $ref: '#/components/responses/Error400'
        '401':
          $ref: '#/components/responses/Error401'
        '404':
          $ref: '#/components/responses/Error404'
        '500':
          $ref: '#/components/responses/Error500'
...
components:
  responses:
    Error400:
      description: '400 Bad Request'
      content:
        application/json:
          schema:
            type: object
            properties:
              error_code:
                type: integer
                example: 400
              message:
                type: string
                example: bad request
    Error401:
      description: '401 Unauthorized'
      content:
        application/json:
          schema:
            type: object
            properties:
              success:
                type: boolean
                example: false
              message:
                type: string
                example: Attempted to perform an unauthorized operation
    Error404:
      description: '404 Not Found'
    Error500:
      description: '500 Internal Server'
      content:
        application/json:
          schema:
            type: object
            properties:
              success:
                type: boolean
                example: false
              message:
                type: string
                example: terjadi kesalahan yang tidak diketahui
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer